requirements.txt
amqp @ file:///home/conda/feedstock_root/build_artifacts/amqp_1703885206855/work
ansi2html @ file:///home/conda/feedstock_root/build_artifacts/ansi2html_1695769064861/work
asttokens==2.4.0
backcall==0.2.0
billiard @ file:///home/conda/feedstock_root/build_artifacts/billiard_1703820024511/work
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1698890160476/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1648883617327/work
celery @ file:///home/conda/feedstock_root/build_artifacts/celery_1703899296064/work
certifi @ file:///croot/certifi_1700501669400/work/certifi
charset-normalizer==3.3.0
click @ file:///home/conda/feedstock_root/build_artifacts/click_1692311806742/work
click-didyoumean @ file:///home/conda/feedstock_root/build_artifacts/click-didyoumean_1642673782881/work
click-plugins==1.1.1
click-repl @ file:///home/conda/feedstock_root/build_artifacts/click-repl_1694959444233/work
comm==0.1.4
dash==2.14.0
dash-auth==1.3.2
dash-bootstrap-components @ file:///home/conda/feedstock_root/build_artifacts/dash-bootstrap-components_1694086999250/work
dash-bootstrap-templates==1.0.5
dash-core-components @ file:///home/conda/feedstock_root/build_artifacts/dash-core-components_1645162191927/work
dash-html-components @ file:///home/conda/feedstock_root/build_artifacts/dash-html-components_1645162180756/work
dash-renderer @ file:///home/conda/feedstock_root/build_artifacts/dash-renderer_1617984936069/work
dash-table @ file:///home/conda/feedstock_root/build_artifacts/dash-table_1645162199872/work
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
Flask==2.2.2
Flask-Compress==1.14
Flask-Login==0.6.3
Flask-SeaSurf==1.1.1
Flask-SQLAlchemy==2.5.1
greenlet==3.0.3
gunicorn==21.2.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1699364556997/work
ipykernel==6.25.2
ipython==8.16.1
itsdangerous @ file:///home/conda/feedstock_root/build_artifacts/itsdangerous_1648147185463/work
jedi==0.19.1
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1654302431367/work
jupyter-dash==0.4.2
jupyter_client==8.4.0
jupyter_core==5.4.0
kombu @ file:///home/conda/feedstock_root/build_artifacts/kombu_1699281730408/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1695367432301/work
matplotlib-inline==0.1.6
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
numpy==1.26.1
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1696202382185/work
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
plotly==5.17.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
retrying==1.3.3
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
SQLAlchemy==1.4.36
stack-data==0.6.3
tenacity @ file:///home/conda/feedstock_root/build_artifacts/tenacity_1692026804430/work
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
ua-parser==0.18.0
urllib3==2.0.6
vine @ file:///home/conda/feedstock_root/build_artifacts/vine_1699264759755/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1700607916581/work
Werkzeug==2.2.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1695255097490/work
